[clang][DFP] Enable target dependent support for decimal floating-point extensions.
Support for decimal floating-point (DFP) extensions will, for most
targets, require run-time library support that, for many targets, is
not yet available. This change adds basic infrastructure support to
enable DFP support to be conditional on the target.

Some targets support more than one possible encoding of DFP values.
It is therefore necessary for a target to not just indicate that
DFP support is available, but also the encoding to be used. This
change recognizes the two known encodings that are used in practice
and ties DFP enablement to the selection of one of these modes.
The known encodings are:
  BID: Binary Integer Decimal
  DPD: Densely Packed Decimal