[SYCL] Initial implementation of SYCLKernelCallStmt and OutlinedFunctionDecl.
SYCLKernelCallStmt is a new Stmt AST node that wraps another substatement
for which it provides a reference to an OutlinedFunctionDecl that provides
a sequence of ImplicitParamDecl and a body Stmt from which a function can
be generated.

Nothing actually constructs a SYCLKernelCallStmt or OutlinedFunctionDecl
at this point; that awaits implementation and integration with an
attribute that nominates a function or function template as representing
a call to a SYCL kernel.