[SYCL] Initial support for the sycl_kernel_entry_point attribute.
The sycl_kernel_entry_point attribute is used to declare a function that
defines a pattern for a SYCL kernel to be emitted. The attribute requires
a single argument that denotes a class type that meets the requirements
for SYCL kernel names as described in section 5.2, "Naming of kernels",
of the SYCL 2020 specification.

This change implements basic support for the sycl_kernel_entry_point
attribute and adds tests for appertainment and handling of its argument list.