[SYCL] Populate the body of OutlinedFunctionDecl AST nodes inserted for sycl_kernel_entry_point attributed functions.
The function body associated with a OutlinedFunctionDecl AST node is a clone
of the original body of the sycl_kernel_entry_point attributed function
modified to substitute references to the original function parameters with
references to replacement variables that stand in for the parameters of
the SYCL kernel caller function that will be emitted during code generation.
This change implements the necessary transforms.