[clang][SYCL] New diagnostic group option: -Wredundant-attribute.
Changes made to issue diagnostics for the sycl_kernel_entry_point attribute
included a new warning for redundant use of the attribute. The new warning
was not associated with a diagnostic group and thus resulted in failure of
test clang/test/AST/ast-dump-sycl-kernel-entry-point.cpp. This change adds
a new -Wredundant-attribute diagnostic group to support diagnosis of
redundant use of attributes generally.