[SYCL] Fix crashing test
Getting the size of parameter requires it to be a complete type.
When the parameter sizes were previously stored in SYCLKernelInfo,
the compiler hit an assert in cases where the definition of the type
was not yet instantiated by the compiler. This PR moves the code
calculating size to the point of invocation of the builtin (instead
of kernel invocation), hoping the type is fully instantiated by the
compiler by then. This needs to be verified by adding new builtin
tests. The existing crashing test now passes because it doesn't
invoke the builtin.