[SYCL] Dummy implementation of code location builtins
The following builtins were implemented to point to the
location of the SYCL kernel object. In the new design,
the SYCL kernel object does not having 'special'
designation since all members of the entry point function
will be treated in a similar manner. It is unclear what
these builtins should now return or whether they should
even exist. This PR is a dummy implementation to make
library compile while we investigate.

__builtin_sycl_kernel_file_name(kernel-name)
__builtin_sycl_kernel_function_name(kernel-name)
__builtin_sycl_kernel_line_number(kernel-name)
__builtin_sycl_kernel_column_number(kernel-name)