[SYCL] Test acceptance of incomplete parameter types in non-defining sycl_kernel_entry_point attributed function declarations.
C++ does not require parameter types to be complete at the point of a
non-defining function declaration. For functions declared with the
sycl_kernel_entry_point attribute, metadata is generated that includes
parameter information. This change adds a test to ensure that incomplete
types are not rejected as parameter types for a non-defining declaration
of a sycl_kernel_entry_point attributed function.