[SYCL] Improve kernel name diagnostic
In case of conflicting SYCL kernel name, the compiler reported
the location of the initial kernel but without any context.
This lead to a poor diagnostic when used with template
as the user had no information on how the initial kernel was instantiated.

A second issue can arise with SFINAE as the name validation was done during
overload resolution when the attribute is used on a templated function.
This lead to the attribute interfering with the overload resolution.

We now record the template instantiation trace when a SYCL entry point is registered.
In case of conflicting name, the diagnostic is issue when emitting the SYCL kernel CallStmt and
the trace is looked up to provide the full context.