[SYCL] Offload kernel code generation for SYCL kernel entry point functions.
A SYCL kernel entry point function is a non-member function or a static member
function declared with the `sycl_kernel_entry_point` attribute. Such functions
define a pattern for an offload kernel to be generated to enable execution of
a SYCL kernel on a device. A SYCL library implementation orchestrates the
invocation of these functions with corresponding SYCL kernel arguments in
response to calls to SYCL kernel invocation functions specified by the SYCL
2020 specification.

SYCL kernel entry point functions are represented in the AST as functions
with a body consisting of a `SYCLKernelCallStmt` and its associated
`OutlinedFunctionDecl` instance. The `OutlinedFunctionDecl` instance specifies
the parameters and body to be used for the generated offload kernel. The
offload kernel is sometimes referred to as a SYCL kernel caller function.

A distinct offload kernel is generated for each SYCL kernel entry point
function defined as a non-inline function or ODR-used in the translation unit.
These offload kernels are given names equivalent to a specialization of a
function template named '__sycl_kernel_caller' with a void return type, no
function parameters, and a single type template argument corresponding to the
type used as the SYCL kernel name as passed as an argument to the
`sycl_kernel_entry_point` attribute. These offload kernels are only emitted
when compiling for a SYCL device target; when compiling for a SYCL host target,
SYCL kernel entry point functions are emitted as normal functions.

Eventually, only those functions that are referenced from an offload kernel
(or declared as SYCL_EXTERNAL) will be emitted when compiling for a device
target. That pruning will be implemented in a future change.