[SYCL] Elizabeth's changes to emit OpenCL kernel metadata for SYCL kernels.
Elizabeth's changes to perform code generation for SYCL kernel device entry
point functions caused the following kernel metadata annotations to be
emitted for all device targets.
- `kernel_arg_addr_space`
- `kernel_arg_access_qual`
- `kernel_arg_type`
- `kernel_arg_base_type`
- `kernel_arg_type_qual`

It turns out that such metadata is not needed for SYCL; the above annotations
are only needed for OpenCL. Investigation revealed that these annotations are
emitted by the oneAPI compiler for ESIMD functions, but that was apparently
due to an incorrect merge conflict resolution that is now being addressed by
https://github.com/intel/llvm/pull/17723.

The above annotations are referenced by the SPIR-V Translator documentation.
See https://github.com/KhronosGroup/SPIRV-LLVM-Translator/blob/main/docs/SPIRVRepresentationInLLVM.rst#function-metadata.

The oneAPI compiler does emit a `kernel_arg_buffer_location` annotation that
is presumed to be needed as part of the support for the dead kernel argument
elimination pass. Support for that annotation does not yet exist upstream.